refactor: split helpers module on error and flags
