+++ src/modules/flags/reducer.js
@@ -6,18 +6,6 @@ import {
   profileOpen,
   profileClose,
 } from "./actions";
-import { loginRequestFailed, registrationRequestFailed } from "../auth";
-import { getCardRequestFailed, postCardRequestFailed } from "../payment";
-
-const error = handleActions(
-  {
-    [loginRequestFailed]: (state, action) => action.payload,
-    [registrationRequestFailed]: (state, action) => action.payload,
-    [getCardRequestFailed]: (state, action) => action.payload,
-    [postCardRequestFailed]: (state, action) => action.payload,
-  },
-  null
-);
 
 const isLoading = handleActions(
   {
@@ -36,7 +24,6 @@ const isProfileOpened = handleActions(
 );
 
 export default combineReducers({
-  error,
   isLoading,
   isProfileOpened,
 });
