+++ src/store/reducer.js
@@ -3,12 +3,14 @@ import auth from "../modules/auth";
 import payment from "../modules/payment";
 import addressesList from "../modules/addressesList";
 import route from "../modules/route";
-import helpers from "../modules/helpers";
+import error from "../modules/error";
+import flags from "../modules/flags";
 
 export default combineReducers({
   auth,
   payment,
   addressesList,
   route,
-  helpers,
+  flags,
+  error,
 });
